Polish the implementation
